// frontend/src/types.ts
export interface TopPlayer {
  id: string
  name: string
  level: number
  association?: string
  power: number
  rank: number
}

export interface TopAssociation {
  id: string
  name: string
  members: number
  power: number
  lingshi: number
  rank: number
}